기능: 레시피 등록 API 구현
* DTO에 Validation 적용
* 수량 필드 타입 변경: int -> double

관련: #23

+++ src/main/java/io/myrecipes/api/domain/RecipeMaterialEntity.java
@@ -18,7 +18,7 @@ public class RecipeMaterialEntity {
     private Integer id;
 
     @Column(nullable = false)
-    private Integer quantity;
+    private Double quantity;
 
     @ManyToOne
     @JsonBackReference
@@ -30,7 +30,7 @@ public class RecipeMaterialEntity {
     private MaterialEntity materialEntity;
 
     @Builder
-    public RecipeMaterialEntity(Integer quantity, RecipeEntity recipeEntity, MaterialEntity materialEntity) {
+    public RecipeMaterialEntity(Double quantity, RecipeEntity recipeEntity, MaterialEntity materialEntity) {
         this.quantity = quantity;
         this.recipeEntity = recipeEntity;
         this.materialEntity = materialEntity;
